// frontend/src/types/about.ts
export interface About {
  _id?: string;

  welcomeTitle: string;
  welcomeSubtitle: string;
  welcomeDescription: string;
  welcomeImage: string;

  aboutTitle: string;
  aboutDescription: string;
  aboutImage: string;

  visionTitle: string;
  visionDescription: string;

  missionTitle: string;
  missionDescription: string;

  buttonText: string;
  buttonLink: string;

  isActive: boolean;
}